### 2024-04-30 #6700 @editorialbot
[PRE REVIEW]: Eaglescope: an interactive visualization and cohort selection tool for biomedical data exploration.
**Submitting author:** @birm (<a href="http://orcid.org/0000-0002-7943-6346">Ryan Birmingham</a>)
**Repository:** https://github.com/sharmalab/eaglescope
**Branch with paper.md** (empty if default branch):
**Version:** 1.0
