### 2024-04-30 6e26db0
add dois except bokeh
