+++ paper.bib
@@ -8,7 +8,8 @@
 @article{cbioportal2013,
   author = {Gao, Jianjiong and Aksoy, Bulent Arman and Dogrusoz, Ugur and Dresdner, Gideon and Gross, Benjamin and Sumer, S Onur and Sun, Yichao and Jacobsen, Anders and Sinha, Rileen and Larsson, Erik and Cerami, Ethan and Sander, Chris and Schultz, Nikolaus},
   title = {Integrative analysis of complex cancer genomics and clinical profiles using the cBioPortal},
-  journal = {Sci Signal},
+  journal = {Science Signaling},
+  doi = {10.1126/scisignal.2004088},
   year = {2013},
   volume = {6},
   number = {269}
@@ -18,6 +19,7 @@
   author = {Iyer, Ganesh and DuttaDuwarah, Sapoonjyoti and Sharma, Ashish},
   title = {DataScope: Interactive visual exploratory dashboards for large multidimensional data},
   booktitle = {2017 IEEE Workshop on Visual Analytics in Healthcare (VAHC)},
+  doi = {10.1109/VAHC.2017.8387496},
   year = {2017},
   pages = {17-23}
 }
@@ -25,6 +27,7 @@
 @article{nbia2020,
   author = {Nguyen, Tin and Shafi, Adib and Nguyen, Tuan-Minh and Schissler, A. Grant and Draghici, Sorin},
   title = {NBIA: a network-based integrative analysis framework - applied to pathway analysis},
+  doi = {10.1038/s41598-020-60981-9},
   journal = {Scientific Reports},
   year = {2020},
   volume = {10},
@@ -34,7 +37,8 @@
 @article{tica2013,
   author = {Clark, Kenneth and Vendt, Bruce and Smith, Kirk and Freymann, John and Kirby, Justin and Moore, Stephen and Phillips, Stanley and Maffit, David and Pringle, Michael and Tarbox, Lawrence and Prior, Fred},
   title = {The Cancer Imaging Archive (TCIA): Maintaining and Operating a Public Information Repository},
-  journal = {J Digit Imaging},
+  journal = {Journal of Digital Imaging},
+  doi = {10.1007/s10278-013-9622-7},
   year = {2013},
   volume = {26},
   number = {6},
@@ -44,6 +48,7 @@
 @article{prism2020,
   author = {Sharma, Ashish and Tarbox, Lawrence and Kurc, Tahsin and Bona, Jonathan and Smith, Kirk and Kathiravelu, Pradeeban and Bremer, Erich and Saltz, Joel H. and Prior, Fred},
   title = {PRISM: A Platform for Imaging in Precision Medicine},
+  doi = {10.1200/cci.20.00001},
   journal = {JCO Clinical Cancer Informatics},
   year = {2020},
   volume = {4}
