+++ paper.bib
@@ -1,8 +1,8 @@
-@misc{bokeh,
-  author = {Bokeh Contributors},
-  title = {Bokeh: an interactive visualization library for modern web browsers},
-  year = {2024},
-  url = {https://github.com/bokeh/bokeh}
+@software{bokeh,
+	title = {Bokeh: an interactive visualization library for modern web browsers},
+	url = {https://github.com/bokeh/bokeh},
+	author = {Bokeh Contributors},
+	date = {2024},
 }
 
 @article{cbioportal2013,
