### 2024-04-30 #6700 @birm
I think we've fixed the dois! Not sure if Bokeh has a preferred citation though.
