### 2024-05-10 b614525
very basic filter count
