+++ dist/BarChart.5f4d3c70.js
@@ -358,7 +358,7 @@ var parent = module.bundle.parent;
 if ((!parent || !parent.isParcelRequire) && typeof WebSocket !== 'undefined') {
   var hostname = "" || location.hostname;
   var protocol = location.protocol === 'https:' ? 'wss' : 'ws';
-  var ws = new WebSocket(protocol + '://' + hostname + ':' + "50047" + '/');
+  var ws = new WebSocket(protocol + '://' + hostname + ':' + "60621" + '/');
   ws.onmessage = function (event) {
     checkedAssets = {};
     assetsToAccept = [];
