### 2024-05-25 #6700 @Kevin-Mattheus-Moerman
@birm thanks for those suggestions. I am currently checking if this work is in scope for JOSS, and if so, will be looking for a handling editor shortly.

In the mean time, can  you help clarify if it would be correct to say that the JavaScript for this submission forms the main code achievement here? Secondly, could you clarify if the JavaScript is fully your original work, or if some default (e.g. GUI related) codes were imported or if code was autogenerated. Thanks.
