### 2024-05-29 #6700 @birm
Sorry for the delay. The code is the accomplishment. It uses react, but it's otherwise the contrubutors' work. The only part which is autogenerated is in ./dist/, which we prebuild for the demo site (https://sharmalab.github.io/eaglescope/) and other deployments where we don't have something to bundle the code.
