### 2024-05-30 #6700 @csoneson
👋🏻 @ianhi, @allie-tatarian - would you be interested in reviewing this submission for JOSS?

> Eaglescope: an interactive visualization and cohort selection tool for biomedical data exploration
https://github.com/openjournals/joss-reviews/issues/6700

Thanks in advance!
