### 2024-05-31 #6700 @allie-tatarian
Hi @csoneson, thanks for reaching out! I actually can't get this software to run on my machine at all - I think it might be a security issue (this has happened to me before). I've attached the error log in case it's useful to the developers (@birm), but I'm going to let someone else review this one.
[2024-05-31T16_28_22_532Z-debug-0.log](https://github.com/user-attachments/files/15516456/2024-05-31T16_28_22_532Z-debug-0.log)
