+++ source/components/VisualTools/VisDataTable/VisDataTableControl/VisSortableItem/VisSortableItem.js
@@ -1,31 +1,42 @@
 import React from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faGripLines } from '@fortawesome/free-solid-svg-icons';
-import { sortableElement, sortableHandle } from 'react-sortable-hoc';
+import { useDrag } from 'react-dnd';
+
 import './VisSortableItem.css';
 
-const DragHandle = sortableHandle(() => (
-  <div className="drag-handle">
-    <FontAwesomeIcon icon={faGripLines} />
-  </div>
-));
+const DragHandle = () => {
+  const [{ isDragging }, drag] = useDrag({
+    collect: (monitor) => ({
+      isDragging: !!monitor.isDragging(),
+    }),
+  });
 
-const VisSortableItem = sortableElement((props) => (
-  <div className="vis-sortable-item ">
-    <DragHandle />
-    <div className="truncated-text" title={props.label}>
-      {props.label}
+  return (
+    <div className="drag-handle" ref={drag}>
+      <FontAwesomeIcon icon={faGripLines} style={{ opacity: isDragging ? 0.4 : 1 }} />
     </div>
+  );
+};
 
-    <div className="input-box">
-      <input
-        type="checkbox"
-        checked={props.isShow}
-        onChange={props.onCheckChanged}
-        value={props.dataKey}
-      />
+const VisSortableItem = (props) => {
+  return (
+    <div className="vis-sortable-item">
+      <DragHandle />
+      <div className="truncated-text" title={props.label}>
+        {props.label}
+      </div>
+
+      <div className="input-box">
+        <input
+          type="checkbox"
+          checked={props.isShow}
+          onChange={props.onCheckChanged}
+          value={props.dataKey}
+        />
+      </div>
     </div>
-  </div>
-));
+  );
+};
 
 export default VisSortableItem;
