+++ VisDataTable.7250dab4.js
@@ -217,10 +217,10 @@ var r=arguments[3];Object.defineProperty(exports,"__esModule",{value:!0}),export
 },{"dom-helpers/contains":"pXI2","prop-types":"D9Od","react":"n8MK","@restart/hooks/useTimeout":"Vil4","warning":"XOCG","uncontrollable":"MTOr","@restart/hooks/useMergedRefs":"omRz","./Overlay":"qQFp","./safeFindDOMNode":"Os8f","react/jsx-runtime":"plwR"}],"gyBy":[function(require,module,exports) {
 
 },{}],"bf3D":[function(require,module,exports) {
-"use strict";Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var e=a(require("react")),t=require("@fortawesome/react-fontawesome"),r=require("@fortawesome/free-solid-svg-icons"),n=require("react-dnd");function a(e){return e&&e.__esModule?e:{default:e}}function l(e,t){return s(e)||c(e,t)||i(e,t)||o()}function o(){throw new TypeError("Invalid attempt to destructure non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}function i(e,t){if(e){if("string"==typeof e)return u(e,t);var r=Object.prototype.toString.call(e).slice(8,-1);return"Object"===r&&e.constructor&&(r=e.constructor.name),"Map"===r||"Set"===r?Array.from(e):"Arguments"===r||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(r)?u(e,t):void 0}}function u(e,t){(null==t||t>e.length)&&(t=e.length);for(var r=0,n=new Array(t);r<t;r++)n[r]=e[r];return n}function c(e,t){var r=null==e?null:"undefined"!=typeof Symbol&&e[Symbol.iterator]||e["@@iterator"];if(null!=r){var n,a,l,o,i=[],u=!0,c=!1;try{if(l=(r=r.call(e)).next,0===t){if(Object(r)!==r)return;u=!1}else for(;!(u=(n=l.call(r)).done)&&(i.push(n.value),i.length!==t);u=!0);}catch(e){c=!0,a=e}finally{try{if(!u&&null!=r.return&&(o=r.return(),Object(o)!==o))return}finally{if(c)throw a}}return i}}function s(e){if(Array.isArray(e))return e}require("./VisSortableItem.css");var f=function(){var a=l((0,n.useDrag)({collect:function(e){return{isDragging:!!e.isDragging()}}}),2),o=a[0].isDragging,i=a[1];return e.default.createElement("div",{className:"drag-handle",ref:i},e.default.createElement(t.FontAwesomeIcon,{icon:r.faGripLines,style:{opacity:o?.4:1}}))},d=function(t){return e.default.createElement("div",{className:"vis-sortable-item"},e.default.createElement(f,null),e.default.createElement("div",{className:"truncated-text",title:t.label},t.label),e.default.createElement("div",{className:"input-box"},e.default.createElement("input",{type:"checkbox",checked:t.isShow,onChange:t.onCheckChanged,value:t.dataKey})))},m=exports.default=d;
[…]
