### 2024-06-04 #6700 @csoneson
👋🏻 @flekschas, @sebastian-raubach - would you be interested in reviewing this submission for JOSS?

> Eaglescope: an interactive visualization and cohort selection tool for biomedical data exploration
https://github.com/openjournals/joss-reviews/issues/6700

Thanks in advance!
