### 2024-06-04 #6700 @sebastian-raubach
> 👋🏻 @flekschas, @sebastian-raubach - would you be interested in reviewing this submission for JOSS?
>
> > Eaglescope: an interactive visualization and cohort selection tool for biomedical data exploration
> > #6700
>
> Thanks in advance!

Yeah, I can give it a go.
