### 2024-06-04 #6837 @editorialbot
[REVIEW]: Eaglescope: an interactive visualization and cohort selection tool for biomedical data exploration.
**Submitting author:** @birm (<a href="http://orcid.org/0000-0002-7943-6346">Ryan Birmingham</a>)
**Repository:** https://github.com/sharmalab/eaglescope
**Branch with paper.md** (empty if default branch):
**Version:** 1.1.1
