### 2024-06-04 #6837 @csoneson
Hi @sebastian-raubach - I'd say that this type of question, which is not strictly related to the functionality or implementation of the software, but rather to the JOSS submission, are better posed in this thread. But there is no strict boundary, and as long as you mention the issues posted in the software repository here, we can keep track of them.

Also, in case it's useful re: your question above, here is a link to the [JOSS authorship policy](https://joss.readthedocs.io/en/latest/submitting.html#authorship).
