### 2024-06-04 sharmalab/eaglescope#118 @sebastian-raubach
Documentation availability
Hello,

I'm trying to find any documentation in addition to installation instructions. the `/docs/` folder just seems to contain a build of the software rather than actual documentation. Is there any documentation that goes along with this tool? Something that explains the configuration file options, a user guide (how to use the interface) etc.

Thanks
