### 2024-06-04 #6837 @sebastian-raubach
Added an issue asking about documentation of the tool https://github.com/sharmalab/eaglescope/issues/118
