### 2024-06-06 #6837 @birm
> @birm Would you be able to clarify the author contribution for me please? Looking at the contribution graphs (https://github.com/sharmalab/eaglescope/graphs/contributors) I can see major code contributions from yourself, Nan Li (with two separate accounts?) and Yahia Zakaria. I can, however, not identify Tony Pan's contribution. Additionally, there are smaller code contributions from other individuals who aren't included in the author list.
>
> If you could shine some light on those two items, that'd be great.
>
> @csoneson As this is my first JOSS review, could you let me know if questions of this nature are best posted on this thread or (as the initial comment suggests) on the software's main repository?
>
> Cheers.

I don't know why Nan used two different accounts, but that's accurate.
Tony Pan is involved with the Eaglescope project in an advisory capacity, especially future planning.
