### 2024-06-07 #6837 @sebastian-raubach
@csoneson I noticed [this section on web-applications](https://joss.readthedocs.io/en/latest/submitting.html#a-note-on-web-based-software) in the JOSS guidelines and I'm not convinced this tool fulfils either of the two requirements. Having said that, just because they used web-technologies, this doesn't necessarily make it a web-application as it doesn't need to be hosted on the web somewhere, you can just run it locally. So I'm a bit unsure about how to proceed. This is further complicated by the fact that I cannot see any form of automated testing in place for this tool which is one of the other checkboxes to tick in the list.

@birm would you be able to confirm whether or not there are automated tests for Eaglescope?
