### 2024-06-12 #6837 @flekschas
Thanks for putting together this tool @birm. It looks really promising! However, I have several remarks:

1. While the `README.md` contains instructions how to set up the development environment. I can't seem to find many details explaining how to configure the dashboard. And according to https://docs.google.com/presentation/d/1zvXCeV-a8k4VercXsgFPTHqml7QwmDDu9snz6dhqIC4/edit#slide=id.g106c66ac588_0_3 the specification file can be quite complex. Moreover, the paper claims that one can "create an interactive dashboard based upon a configuration file **and either an API or data file**". How is one supposed to do the later? Unless I'm missing something, the `README.md` does speak to any of that. So at the moment I can't tick `Example usage` as I really don't know how to set up the tool myself.
2. Similarly, while the repo contains a live demo, it's somewhat unclear to me what all the visualizations do and how they work together to solve some scientific questions. The video gets at that to some degree but it's a) 3 years old and b) only one minute long (and hence doesn't really cover much). There needs to be a bit more for me to check `Functionality documentation`
3. As far as I can tell, there are no tests of any kind
4. There are also no community guidelines
6. While the authors provide a statement of need, the paper is currently lacking a section on how it compares to other charting libraries. (Hence I cannot check `State of the field`) Part of the issue here is also that I'm not clear on what the exact scope of this tool is. In the very first sentence the authors say "... and cohort selection tool designed for biomedical data exploration" and it appears as if this tool is designed to visualize cohort data. But later on cohort data is not mentioned anymore and instead the authors talk about "exploring large biomedical datasets". Unfortunately I don't believe that this tool works with all kinds of biomedical. In particular, I doubt that the tool supports large datasets given that we're talking about a static HTML web app that uses SVG for rendering. (One cannot render millions or billions of data points with SVG as are commonly found in genomics or single-cell biology.) I'm not bringing this up to say the tool isn't useful but I think the paper needs to be revised to be more specific as to what types of biomedical data are actually supported. Having done that it should also be easy to relate the tool to all the other biomedical visualization tools that are out there. (Speaking from the downloaded wine dataset, my current assumption is that this tool can explore small tabular data. I'm happy to be convinced otherwise but the working demos all point to small datasets.)

In summary, this truly looks like a neat tool but it (primarily) needs documentation and the exact use case its supports should be worked out more in the paper.

PS: After digging around in the code I found out how one can change the config. The wine, demo, and collection-vis configs work but it seems that `clinical-vis-config.json` and `vis-config.json` are broken (I only see a loading spinner). It's odd though to require an end-user to dig into the source code to specify the default config that's being loaded.
