### 2024-06-12 c5af76f
initial test, roughly
