+++ jest.config.js
@@ -1,0 +1,13 @@
+module.exports = {
+    testEnvironment: 'jsdom',
+    moduleFileExtensions: ['js', 'jsx'],
+    transform: {
+      '^.+\\.jsx?$': 'babel-jest',
+    },
+    moduleNameMapper: {
+      '\\.css$': 'identity-obj-proxy',
+    },
+    setupFilesAfterEnv: ['@testing-library/jest-dom'],
+    setupFilesAfterEnv: ['./tests/setupTests.js'],
+  };
+  
