+++ tests/setupTests.js
@@ -1,0 +1,3 @@
+const fetch = require('cross-fetch');
+
+global.fetch = fetch;
