### 2024-06-12 c0d4ad3
fixed test cases for config, title, vis
