### 2024-06-12 b5e9d76
lint fix, allow objects for this, defaults
