### 2024-06-12 9647fb2
Update Readme with more descriptive use instructions
