+++ README.md
@@ -3,47 +3,179 @@
 [<h2 align="center"><img src="./eaglescope.png" width="200" /></h2>](.eaglescope.png)
 
 
-Eaglescope is a configurable code-free interactive visualization and cohort selection tool designed for biomedical data exploration. 
+Eaglescope is a configurable code-free interactive visualization and cohort selection tool designed for biomedical data exploration. It is designed to be hosted flexibly without the need for a dedicated server, and creates an interactive dashboard based upon a configuration file and either an API or data file. It uses visualizations of sets of features to describe and enable contextual filtering of the data. This allows for users to understand deeper patterns or anomalies within the data, and to create datasets specifically tuned to their requirements effortlessly. Eaglescope is typically utilized either as a tool to create refined datasets tailored for training and validating machine learning AI models, or as a central hub for further exploration, allowing users to seamlessly navigate to biomedical viewers such as DICOM or whole slide imaging (WSI) platforms. 
+
+![Interactive Contextual Visualizations](./ContextualVis.png)
 
 ## Development
-Install dependencies using `npm install`
+Install dependencies by running `npm install`
 
-Develop mode using `npm run dev`
+Use develop mode by running `npm run dev`
 
-
-### Running ###
-
-
-##### Prerequisites
-
-* Install [Node.js](https://nodejs.org/en/download/) and [NPM](https://www.npmjs.com/get-npm)
+Build ./dist for use with a static web server by running `npm run build`
 
 
-##### Installation
+# Setup and Configuration
 
-* Clone the [repository](https://github.com/birm/datascope2.git)
-* Enter the eaglescope directory (this directory)
-* Get dependencies with ```npm install```
-* Run ```npm run build```
+The simplest way to set up eaglescope is to build (`npm run build`) then host the contents of ./dist/ using a static web server such as rapache or `python -m http.server`. Additionally, configuration and data resources are already on the web with appropriate cross origin headers, you can simply navigate to [https://sharmalab.github.io/eaglescope/?configurl=](https://sharmalab.github.io/eaglescope/?configurl=)(configuration json file).
 
-##### Running
-* Modify the files present in ```./config/vis-config.json``` to fit your needs:
-    * UNIT_OF_GRID_VIEW                   (For basic unit in each viusal component)
-    * MARGIN_OF_GRID_VIEW                 (For setting the margin unit in pixel) 
-    * DATA_RESOURCE_URL                   (For data resource)
-    * VISUALIZATION_VIEW_CONFIGURATION    (For dashboard settings)
+Each Eaglescope dashboard needs a configuration url which contains global information for the dashboard, information about the data resource and how to read it, and configuration for each visualization.
 
 
-* Run ```npm run dev```
-* Go to ```http://localhost:1234```.
+<table>
+  <tr>
+   <td><strong>CONFIGURATION FIELD</strong>
+   </td>
+   <td><strong>DESCRIPTION</strong>
+   </td>
+   <td><strong>EXAMPLE VALUE</strong>
+   </td>
+  </tr>
+  <tr>
+   <td>HOME_URL
+   </td>
[…]
