### 2024-06-12 #6837 @birm
We've added community guidelines, improved the documentation, and added (slightly) better testing for eaglescope.

I think we still need to address points 2 and 5 @flekschas
