### 2024-06-13 #6837 @csoneson
> @csoneson I noticed [this section on web-applications](https://joss.readthedocs.io/en/latest/submitting.html#a-note-on-web-based-software) in the JOSS guidelines and I'm not convinced this tool fulfils either of the two requirements. Having said that, just because they used web-technologies, this doesn't necessarily make it a web-application as it doesn't need to be hosted on the web somewhere, you can just run it locally. So I'm a bit unsure about how to proceed. This is further complicated by the fact that I cannot see any form of automated testing in place for this tool which is one of the other checkboxes to tick in the list.

@sebastian-raubach thanks for your comment and sorry for the delayed response, I was out of office for a few days. We have discussed it in the editor team and we do feel that this is technically in scope - however, as you (and @flekschas) have pointed out, implementing sufficient testing to be able to catch unexpected behaviour in a way that is as automated as possible will be essential.

@birm - I would also suggest providing direct links to tests, contribution guidelines and similar in the README, so that they are easy to find.
