### 2024-06-13 0a131c7
update data table selection logic
