### 2024-06-15 2c849e0
cache csv files
they're often slow/big and don't change a ton. Maybe make this a setting instead?
