+++ source/common/utils.js
@@ -1,55 +1,4 @@
-// chart = {
-//   id: UniqueKey.WITH_CNA_DATA,
-//   title: "With CNA Data",
-//   description: "With CNA Data",
-//   chartType: ChartTypeEnum.PIE_CHART,
-//   //dataType: ChartMetaDataTypeEnum.GENOMIC,
-//   //patientAttribute: false,
-//   size: [w, h], // size [width, height]
-//   priority: 0 //
-//   //renderWhenDataChange: false
-// };
-
-// layout = {
-//   i: "key",
-//   x: "x",
-//   y: "y",
-//   w: "width",
-//   h: "height",
-//   isResizable: false
-// };
-
-
-export function numFixed(num, digits = 2) {
-  return Number.isInteger(num) ? num : num.toFixed(digits);
-}
-export function isEquivalent(a, b) {
-  // Create arrays of property names
-  const aProps = Object.getOwnPropertyNames(a);
-  const bProps = Object.getOwnPropertyNames(b);
-
-  // If number of properties is different,
-  // objects are not equivalent
-  if (aProps.length !== bProps.length) {
-    return false;
-  }
-
-  for (let i = 0; i < aProps.length; i++) {
-    const propName = aProps[i];
-
-    // If values of same property are not equal,
-    // objects are not equivalent
-    if (a[propName] !== b[propName]) {
-      return false;
-    }
-  }
-
-  // If we made it this far, objects
-  // are considered equivalent
-  return true;
-}
-
-/// create a matrix
+// Create a matrix
 export function createMatrix(rows, cols = rows) {
   const matrix = [];
   for (let i = 0; i < rows; i++) {
@@ -58,64 +7,92 @@ export function createMatrix(rows, cols = rows) {
   return matrix;
 }
 
-// has Space
+// Check if there is space in the matrix at the given position for the given size
 export function hasSpace(matrix, pos, size) {
   if (matrix.length === 0) return true;
+  let isHasSpace = true;
   for (let i = pos[1]; i < pos[1] + size[1]; i++) {
-    if (i >= matrix.length) return false;
+    if (i >= matrix.length) {
+      break;
+    }
     for (let j = pos[0]; j < pos[0] + size[0]; j++) {
-      if (j >= matrix[0].length || matrix[i][j] !== null) return false;
+      if (j >= matrix[0].length || matrix[i][j]) {
+        isHasSpace = false;
+        break;
+      }
+    }
+    if (!isHasSpace) {
+      break;
     }
   }
-  return true;
+  return isHasSpace;
 }
 
+// Get the position for the chart in the matrix
+function getPosition(matrix, size) {
+  let position;
+
+  // Try to fit in existing rows first
+  for (let i = 0; i < matrix.length; i++) {
+    if (matrix[i].some(cell => cell !== null)) { // Check if the row is not empty
+      for (let j = 0; j < matrix[i].length; j++) {
+        if (matrix[i][j] === null && hasSpace(matrix, [j, i], size)) {
+          return [j, i];
+        }
+      }
+    }
+  }
+
+  // If no suitable position is found, create a new row
+  return [0, matrix.length];
+}
+
+// Fill the matrix with the chart id at the given position and size
 export function fillMatrix(matrix, val, pos = [0, 0], size = [matrix[0].length, matrix.length]) {
   for (let i = pos[1]; i < pos[1] + size[1]; i++) {
+    while (matrix.length <= i) {
+      matrix.push(new Array(matrix[0].length).fill(null));
+    }
     for (let j = pos[0]; j < pos[0] + size[0]; j++) {
       matrix[i][j] = val;
     }
   }
 }
 
[…]
