### 2024-06-17 3b10616
hold off on allow sort after resize
