+++ source/components/VisualTools/VisDataTable/SelectDataTable.js
@@ -197,6 +197,14 @@ export default class SelectDataTable extends PureComponent {
           onDrag={(event, { deltaX }) => {
             this.resizeRow({ dataKey, deltaX });
           }}
+          onStart={(event, { deltaX }) => {
+            this.setState({"isResize": true});
+          }}
+          onStop={(event, { deltaX }) => {
+            setTimeout(() => {
+              this.setState({"isResize": false})
+            }, 300);  // 300 milliseconds delay
+          }}
           position={{ x: 0 }}
           zIndex={999}
         >
@@ -231,7 +239,10 @@ export default class SelectDataTable extends PureComponent {
   }
 
   sortHandler({ sortBy, sortDirection }) {
-    this.setState({ sortBy, sortDirection });
+
+    if (!this.state.isResize){
+      this.setState({ sortBy, sortDirection });
+    }
   }
 
   render() {
