### 2024-06-17 b353089
rebuild select table
