+++ dist/KMCurve.b8405d3d.js
@@ -382,7 +382,7 @@ var parent = module.bundle.parent;
 if ((!parent || !parent.isParcelRequire) && typeof WebSocket !== 'undefined') {
   var hostname = "" || location.hostname;
   var protocol = location.protocol === 'https:' ? 'wss' : 'ws';
-  var ws = new WebSocket(protocol + '://' + hostname + ':' + "51667" + '/');
+  var ws = new WebSocket(protocol + '://' + hostname + ':' + "52833" + '/');
   ws.onmessage = function (event) {
     checkedAssets = {};
     assetsToAccept = [];
