+++ source/hooks/useFetch.js
@@ -4,7 +4,7 @@ import * as d3 from 'd3';
 function isNumeric(str) {
   if (typeof str !== 'string') return false; // we only process strings!
   return (
-    !isNaN(str)
+    !Number.isNaN(str)
     // use type coercion to parse the _entirety_ of the string
     // (`parseFloat` alone does not do this)...
     && !Number.isNaN(parseFloat(str))
