+++ dist/index.js
@@ -193,7 +193,7 @@ var parent = module.bundle.parent;
 if ((!parent || !parent.isParcelRequire) && typeof WebSocket !== 'undefined') {
   var hostname = "" || location.hostname;
   var protocol = location.protocol === 'https:' ? 'wss' : 'ws';
-  var ws = new WebSocket(protocol + '://' + hostname + ':' + "52833" + '/');
+  var ws = new WebSocket(protocol + '://' + hostname + ':' + "58514" + '/');
   ws.onmessage = function (event) {
     checkedAssets = {};
     assetsToAccept = [];
