### 2024-06-17 1f78ade
use intl collator sort for bar charts
