+++ source/components/VisualTools/Chart/HorizontalBarChart.js
@@ -24,10 +24,14 @@ const transform = (data, field, isList = false) => {
   if (isList) {
     return transformList(data, field);
   }
+  const collator = new Intl.Collator(undefined, { numeric: true, sensitivity: 'base' });
+  function collSort(a,b){
+    return collator.compare(a,b);
+  }
   return d3
     .nest()
     .key((d) => d[field])
-    .sortKeys(d3.ascending)
+    .sortKeys(collSort)
     .rollup((v) => v.length)
     .entries(data);
 };
