+++ source/components/VisualTools/VisDataTable/SelectDataTable.js
@@ -148,15 +148,16 @@ export default class SelectDataTable extends PureComponent {
   }
 
   getSortData() {
+    const collator = new Intl.Collator(undefined, { numeric: true, sensitivity: 'base' });
     const { data, filterData, filters } = this.props;
     const { sortBy, sortDirection } = this.state;
     const currentData = filters.length > 0 ? filterData : data;
-    // filter TODO
+
     return sortBy && sortDirection
       ? currentData.sort((a, b) => {
         const first = sortDirection === SortDirection.ASC ? a : b;
         const second = sortDirection === SortDirection.ASC ? b : a;
-        return first[sortBy] > second[sortBy] ? 1 : -1;
+        return collator.compare(first[sortBy], second[sortBy]);
       })
       : currentData;
   }
