### 2024-06-18 6da58f0
stop prop on drag handle doe vistable
