### 2024-06-18 63bdafc
copy delay state over
