+++ dist/HorizontalBarChart.fcfb2c72.js
@@ -166,7 +166,7 @@ var transform = function transform(data, field) {
     sensitivity: 'base'
   });
   function collSort(a, b) {
-    return collator.compare(a, b);
+    return collator.compare(b, a);
   }
   return d3.nest().key(function (d) {
     return d[field];
@@ -357,7 +357,7 @@ var parent = module.bundle.parent;
 if ((!parent || !parent.isParcelRequire) && typeof WebSocket !== 'undefined') {
   var hostname = "" || location.hostname;
   var protocol = location.protocol === 'https:' ? 'wss' : 'ws';
-  var ws = new WebSocket(protocol + '://' + hostname + ':' + "51744" + '/');
+  var ws = new WebSocket(protocol + '://' + hostname + ':' + "52132" + '/');
   ws.onmessage = function (event) {
     checkedAssets = {};
     assetsToAccept = [];
