### 2024-06-21 b5a74f6
add code of conduct and tests
