### 2024-06-27 0ced388
color checkboxes once downloaded
