### 2024-06-27 a93b38a
safer and working selection dl logic
