### 2024-06-27 d4a07b3
modify claims about eaglescope to focus on tabular/cohort
