### 2024-07-01 #6837 @sebastian-raubach
> 👋🏻 Just wanted to check in to see where things are at here. Let me know if you have any questions. Thanks!

Hi Charlotte, for me things are kind of stuck on the automated testing criterion. I can see that there has been a little work done on that front but ultimately, only the smoke test (does it even compile and run?) and some very limited other test (does it show the correct page title and show one visualization element) are included. There are no tests that look into the functionality of the application, e.g. is the input data loaded correctly? does the filtering work? are interactions with the charts working as expected? is the exported configuration correct? None of these are covered.

Since you mentioned in one of your earlier replies  the editorial team wanted to ensure that things like this are covered, which at this point they aren't.

There is also more that could be done on the documentation front. I am happy to see that the format and structure of the configuration files has been added as documentation, but there is no user-facing documentation when it comes to the use of the interface.

Finally, there is no "state of the field" section to speak of where the tool would be compared and evaluated against other tools in the area.
