### 2024-07-17 cc59d15
better marked selected downloads
