### 2024-07-18 #6837 @flekschas
I had another look at the repo and the paper and largely agree with @sebastian-raubach.

I cannot check "Functionality: Have the functional claims of the software been confirmed?" because there are not test verifying the functionality. The smoke test is a great start but it really isn't anything other than a start. If there was a basic test for each chart type I'd say it's okay.

Similarly, while very basic documentation is now available, I'd still have to dig into the source code to understand how to configure examples. Similarly, community guidelines like a short guide for how to contribute and issue/pr templates are not present. I checked "Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems)." because the demo page does provide an example and there are more under "config". However, it'd be nice if there was some sort of description added to the demo page explaining what data one is looking at. (I know it's the classic wine dataset but other visitors might not).

Finally, the paper edits are step in the right direction but a comparison to other similar visualization software and papers for cohort visualization and dashboarding is still missing.

@birm could you take a look at each missing check box and ping us once you feel like you have addressed them?
@csoneson I appreciate the progress the authors have made but there's quite a bit of work left to do.
