### 2024-07-18 1aa2c0f
add figure role to charts for a11y
