+++ source/components/VisualTools/Chart/HorizontalBarChart.js
@@ -186,7 +186,7 @@ function HorizontalBarChart(props) {
     }, 100);
   }, [props.filters, props.filterData, props.layout]);
 
-  return <div id={props.id} ref={self} style={{ width: '100%', height: '100%' }} />;
+  return <div id={props.id} ref={self} role="figure" style={{ width: '100%', height: '100%' }} />;
 }
 
 export default HorizontalBarChart;
