### 2024-07-18 e5070bc
add one test per chart type
