### 2024-07-18 #6837 @birm
As a partial update, I've been working on this. I've added more tests in https://github.com/sharmalab/eaglescope/pull/127, but I suspect we could use quite a few more still.

I'm aware that I have a lot to cover here, and I appreciate the guidance and patience.
