### 2024-07-19 b513dca
Update and rename smoke_test.yml to jest_tests.yml
