### 2024-08-05 6dd1b6a
add log scale for bar charts
