### 2024-08-05 9b78c87
build log scale bar charts
