### 2024-08-19 52a6c35
get just ext in most cases
