### 2024-08-19 fba8070
build with ext split
