### 2024-09-12 7761ec1
clarify readme, config
