### 2024-09-12 e3afa9a
override image squeeze
