### 2024-09-12 49fbd41
refactor table in hopes it helps
