+++ paper.bib
@@ -1,55 +1,127 @@
 @software{bokeh,
-	title = {Bokeh: an interactive visualization library for modern web browsers},
-	url = {https://github.com/bokeh/bokeh},
-	author = {Bokeh Contributors},
-	date = {2024},
+  title = {Bokeh: An Interactive Visualization Library for Modern Web Browsers},
+  url = {https://github.com/bokeh/bokeh},
+  author = {Bokeh Contributors},
+  date = {2024}
 }
 
 @article{cbioportal2013,
   author = {Gao, Jianjiong and Aksoy, Bulent Arman and Dogrusoz, Ugur and Dresdner, Gideon and Gross, Benjamin and Sumer, S Onur and Sun, Yichao and Jacobsen, Anders and Sinha, Rileen and Larsson, Erik and Cerami, Ethan and Sander, Chris and Schultz, Nikolaus},
-  title = {Integrative analysis of complex cancer genomics and clinical profiles using the cBioPortal},
+  title = {Integrative Analysis of Complex Cancer Genomics and Clinical Profiles Using the cBioPortal},
   journal = {Science Signaling},
-  doi = {10.1126/scisignal.2004088},
   year = {2013},
   volume = {6},
-  number = {269}
+  number = {269},
+  doi = {10.1126/scisignal.2004088}
 }
 
 @inproceedings{datascope2017,
   author = {Iyer, Ganesh and DuttaDuwarah, Sapoonjyoti and Sharma, Ashish},
-  title = {DataScope: Interactive visual exploratory dashboards for large multidimensional data},
+  title = {DataScope: Interactive Visual Exploratory Dashboards for Large Multidimensional Data},
   booktitle = {2017 IEEE Workshop on Visual Analytics in Healthcare (VAHC)},
-  doi = {10.1109/VAHC.2017.8387496},
   year = {2017},
-  pages = {17-23}
+  pages = {17--23},
+  doi = {10.1109/VAHC.2017.8387496}
 }
 
 @article{nbia2020,
   author = {Nguyen, Tin and Shafi, Adib and Nguyen, Tuan-Minh and Schissler, A. Grant and Draghici, Sorin},
-  title = {NBIA: a network-based integrative analysis framework - applied to pathway analysis},
-  doi = {10.1038/s41598-020-60981-9},
+  title = {NBIA: A Network-Based Integrative Analysis Framework - Applied to Pathway Analysis},
   journal = {Scientific Reports},
   year = {2020},
   volume = {10},
-  number = {1}
+  number = {1},
+  doi = {10.1038/s41598-020-60981-9}
 }
 
 @article{tica2013,
   author = {Clark, Kenneth and Vendt, Bruce and Smith, Kirk and Freymann, John and Kirby, Justin and Moore, Stephen and Phillips, Stanley and Maffit, David and Pringle, Michael and Tarbox, Lawrence and Prior, Fred},
   title = {The Cancer Imaging Archive (TCIA): Maintaining and Operating a Public Information Repository},
   journal = {Journal of Digital Imaging},
-  doi = {10.1007/s10278-013-9622-7},
   year = {2013},
   volume = {26},
   number = {6},
-  pages = {1045-1057}
+  pages = {1045--1057},
+  doi = {10.1007/s10278-013-9622-7}
 }
 
 @article{prism2020,
   author = {Sharma, Ashish and Tarbox, Lawrence and Kurc, Tahsin and Bona, Jonathan and Smith, Kirk and Kathiravelu, Pradeeban and Bremer, Erich and Saltz, Joel H. and Prior, Fred},
[…]
