### 2024-09-12 #6837 @birm
Okay, I believe I've addressed most of the comments in some form between (tests) https://github.com/sharmalab/eaglescope/pull/127 and (paper/readme) https://github.com/sharmalab/eaglescope/pull/130. Thanks for the nudge to make this happen! Looking forward to more feedback. And once again, thanks for the patience.
