### 2024-09-23 #6837 @sebastian-raubach
Ok, I've had a chance to look over the proposed changes. Below are some comments, suggestions and questions.

I can see you added more tests, specifically, you added tests for each chart type. I think this is a good idea. I do, however, have a question regarding what the test actually checks. You're creating a chart with some very basic mock data, then after the chart is created, you check `expect(document.getElementById(mockId)).toBeInTheDocument();` which (I assume) checks whether the given chart has been added to the HTML document. My question here is: Would this check ever fail? I.e. if the chart is created with invalid data, will `.toBeInDocument()` actually return a falsy value?

Ideally, I'd like to see more extensive testing of the functionality of each component and their connectivity via the filters, but I realise that this is a very involved task.

Manuscript line 28 should be "where the criteria are known".

Apart from these, I think the modifications made to the tool documentation and the manuscript improve the overall quality significantly.
