### 2024-09-25 #6837 @flekschas
It'd really go a long way if the authors could lay out how they think they've addressed the concerns instead of having us to dig into the source code.

From my previous review:

> I cannot check "Functionality: Have the functional claims of the software been confirmed?" because there are no tests verifying the functionality. The smoke test is a great start but it really isn't anything other than a start. If there was a basic test for each chart type I'd say it's appropriately tested.

Looks like each chart type now has an existence test associated to it. That's great! :tada: However, as @sebastian-raubach pointed out, it doesn't really test the functionality. Unless there's a unexpected runtime error, the test will always pass if I'm not mistaken. A very simply but effective approach to verifying that the components work as expected (and produce a chart of some type) would be to turn the HTML document into an image and compare it to a static assets. That shouldn't be hard to do with a library like https://www.npmjs.com/package/html-to-image. One could also do other functionality checks like verifying that a pie chart that should have 3 slices indeed has three SVG path elements or so.

Hence, I cannot check "**Automated tests**: Are there automated tests or manual steps described so that the functionality of the software can be verified?".

> Similarly, while very basic documentation is now available, I still have to dig into the source code to understand how to configure each visualization.

The added table is acceptable documentation. Hence, I checked "Functionality documentation".

> Community guidelines like a short "how to contribute guide" and issue/pr templates are not present

No news on this front as far as I know. Hence, I cannot check "Community guidelines".

> Finally, the paper edits are a step in the right direction but a comparison to other similar visualization software and papers for cohort visualization and dashboarding is still missing.

The authors added the missing section. Hence, I checked "State of the field".
