### 2024-09-27 6fa6f16
wait and check test contents
