+++ tests/Heatmap.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { render } from '@testing-library/react';
+import { render, screen, waitFor } from '@testing-library/react';
 import Heatmap from '../source/components/VisualTools/Chart/Heatmap';
 import '@testing-library/jest-dom';
 
@@ -24,18 +24,28 @@ const mockLayout = {
   currentCols: 2,
 };
 
-test('renders without crashing', () => {
-  render(
-    <Heatmap
-      data={mockData}
-      fields={mockFields}
-      id={mockId}
-      title={mockTitle}
-      filterData={mockData}
-      filters={mockFilters}
-      filterAdded={mockFilterAdded}
-      layout={mockLayout}
-    />
-  );
-  expect(document.getElementById(mockId)).toBeInTheDocument();
+describe('Heatmap Vis Component', () => {
+  test('renders box contents', async() => {
+    const {container} = render(
+      <Heatmap
+        data={mockData}
+        fields={mockFields}
+        id={mockId}
+        title={mockTitle}
+        filterData={mockData}
+        filters={mockFilters}
+        filterAdded={mockFilterAdded}
+        layout={mockLayout}
+      />
+    );
+
+    await waitFor(() => {
+      const chartElement = screen.getByRole('figure', { hidden: true });
+      // console.log(chartElement.innerHTML);
+      expect(chartElement).toBeInTheDocument();
+      // nonsense test for now
+      const rectElements = chartElement.querySelectorAll('rect');
+      expect(rectElements).toHaveLength(4); // comparison boxes
+    }, { timeout: 3000 });
+  });
 });
