+++ paper.md
@@ -43,7 +43,7 @@ To create a dashboard, users simply need to create a file specifying the data so
 
 # State of the Field
 
-Cohort definition is an essential aspect of both retrospective and prospective studies. Different workflows are used to define cohorts depending on the nature and needs of a study. For example, in cases where the criteria is known, retrospective data for the criteria can be downloaded using user-focused natural language tools such as Criteria2Query [@Yuan2019]. 
+Cohort definition is an essential aspect of both retrospective and prospective studies. Different workflows are used to define cohorts depending on the nature and needs of a study. For example, in cases where the criteria are known, retrospective data for the criteria can be downloaded using user-focused natural language tools such as Criteria2Query [@Yuan2019]. 
 
 In addition to statistical and code methods, visual tools can be and are used in cohort studies. A tool called Composer [@Rogers2019-ya] allows for surgeons and patients to visually interpret results of cohort studies. Some tools, such as CAVA (Cohort Analysis via Visual Analytics) [@Zhang2015-kd] can be used for both batch analytics and interactive analytics for exploratory cohort analysis for data stored in its own relational database.
 
