### 2024-09-28 #6837 @birm
Thank you both for nudging me to make the tests better; we've done https://github.com/sharmalab/eaglescope/pull/134 which doesn't test everything, but it at least checks generally that the elements expected within the visualization are present. The exception is the "ParallelCoordinates" chart which only uses a canvas; so far I'm just checking that there is indeed a canvas which isn't a very good test. A good next iteration of these tests should check how the rendered element actually looks using image comparison, but I have not implemented this yet. Currently it is almost entirely tests like "has the right number of pie chart slices/bars" and "renders the axes" etc.

We've also fixed the typo @sebastian-raubach pointed out. Thank you for your vigilance.

We've also added some basic templates and another copy of the creator covenant in https://github.com/sharmalab/.github
