### 2024-10-10 2dc5f5c
memoize and change one case to default
