### 2024-10-10 b482c84
add option for split view
