+++ BarChart.fa4140c9.js
@@ -1,4 +1,0 @@
-parcelRequire=function(e,r,t,n){var i,o="function"==typeof parcelRequire&&parcelRequire,u="function"==typeof require&&require;function f(t,n){if(!r[t]){if(!e[t]){var i="function"==typeof parcelRequire&&parcelRequire;if(!n&&i)return i(t,!0);if(o)return o(t,!0);if(u&&"string"==typeof t)return u(t);var c=new Error("Cannot find module '"+t+"'");throw c.code="MODULE_NOT_FOUND",c}p.resolve=function(r){return e[t][1][r]||r},p.cache={};var l=r[t]=new f.Module(t);e[t][0].call(l.exports,p,l,l.exports,this)}return r[t].exports;function p(e){return f(p.resolve(e))}}f.isParcelRequire=!0,f.Module=function(e){this.id=e,this.bundle=f,this.exports={}},f.modules=e,f.cache=r,f.parent=o,f.register=function(r,t){e[r]=[function(e,r){r.exports=t},{}]};for(var c=0;c<t.length;c++)try{f(t[c])}catch(e){i||(i=e)}if(t.length){var l=f(t[t.length-1]);"object"==typeof exports&&"undefined"!=typeof module?module.exports=l:"function"==typeof define&&define.amd?define(function(){return l}):n&&(this[n]=l)}if(parcelRequire=f,i)throw i;return f}({"KFlv":[function(require,module,exports) {
-"use strict";Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var t=u(require("react")),e=u(require("d3")),r=a(require("prop-types")),n=a(require("../../partials/tooltip"));function a(t){return t&&t.__esModule?t:{default:t}}function i(t){if("function"!=typeof WeakMap)return null;var e=new WeakMap,r=new WeakMap;return(i=function(t){return t?r:e})(t)}function u(t,e){if(!e&&t&&t.__esModule)return t;if(null===t||"object"!=typeof t&&"function"!=typeof t)return{default:t};var r=i(e);if(r&&r.has(t))return r.get(t);var n={__proto__:null},a=Object.defineProperty&&Object.getOwnPropertyDescriptor;for(var u in t)if("default"!==u&&{}.hasOwnProperty.call(t,u)){var o=a?Object.getOwnPropertyDescriptor(t,u):null;o&&(o.get||o.set)?Object.defineProperty(n,u,o):n[u]=t[u]}return n.default=t,r&&r.set(t,n),n}var o=function(t,e){var r=new Map;return t.forEach(function(t){var n=t[e];Array.isArray(n)?n.forEach(function(t){r.has(t)||r.set(t,0),r.set(t,r.get(t)+1)}):(r.has(n)||r.set(n,0),r.set(n,r.get(n)+1))}),Array.from(r).map(function(t){return{key:t[0],value:t[1]}})},l=function(t,r){return arguments.length>2&&void 0!==arguments[2]&&arguments[2]?o(t,r):e.nest().key(function(t){return t[r]}).sortKeys(e.ascending).rollup(function(t){return t.length}).entries(t)},s=function(t,r){t.each(function(){var t,n=e.select(this),a=n.text().split(/\s+/).reverse(),i=[],u=0,o=n.attr("y"),l=parseFloat(n.attr("dy")),s=n.text(null).append("tspan").attr("x",0).attr("y",o).attr("dy","".concat(l,"em"));for(t=a.pop();t;)i.push(t),s.text(i.join(" ")),s.node().getComputedTextLength()>r&&(i.pop(),s.text(i.join(" ")),i=[t],s=n.append("tspan").attr("x",0).attr("y",o).attr
[…]
