+++ HorizontalBarChart.f6343758.js
@@ -1,4 +1,0 @@
-parcelRequire=function(e,r,t,n){var i,o="function"==typeof parcelRequire&&parcelRequire,u="function"==typeof require&&require;function f(t,n){if(!r[t]){if(!e[t]){var i="function"==typeof parcelRequire&&parcelRequire;if(!n&&i)return i(t,!0);if(o)return o(t,!0);if(u&&"string"==typeof t)return u(t);var c=new Error("Cannot find module '"+t+"'");throw c.code="MODULE_NOT_FOUND",c}p.resolve=function(r){return e[t][1][r]||r},p.cache={};var l=r[t]=new f.Module(t);e[t][0].call(l.exports,p,l,l.exports,this)}return r[t].exports;function p(e){return f(p.resolve(e))}}f.isParcelRequire=!0,f.Module=function(e){this.id=e,this.bundle=f,this.exports={}},f.modules=e,f.cache=r,f.parent=o,f.register=function(r,t){e[r]=[function(e,r){r.exports=t},{}]};for(var c=0;c<t.length;c++)try{f(t[c])}catch(e){i||(i=e)}if(t.length){var l=f(t[t.length-1]);"object"==typeof exports&&"undefined"!=typeof module?module.exports=l:"function"==typeof define&&define.amd?define(function(){return l}):n&&(this[n]=l)}if(parcelRequire=f,i)throw i;return f}({"UEE2":[function(require,module,exports) {
-"use strict";Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var e=u(require("react")),t=u(require("d3")),r=i(require("prop-types")),n=i(require("../../partials/tooltip"));function i(e){return e&&e.__esModule?e:{default:e}}function a(e){if("function"!=typeof WeakMap)return null;var t=new WeakMap,r=new WeakMap;return(a=function(e){return e?r:t})(e)}function u(e,t){if(!t&&e&&e.__esModule)return e;if(null===e||"object"!=typeof e&&"function"!=typeof e)return{default:e};var r=a(t);if(r&&r.has(e))return r.get(e);var n={__proto__:null},i=Object.defineProperty&&Object.getOwnPropertyDescriptor;for(var u in e)if("default"!==u&&{}.hasOwnProperty.call(e,u)){var l=i?Object.getOwnPropertyDescriptor(e,u):null;l&&(l.get||l.set)?Object.defineProperty(n,u,l):n[u]=e[u]}return n.default=e,r&&r.set(e,n),n}var l=function(e,t){var r=new Map;return e.forEach(function(e){var n=e[t];Array.isArray(n)?n.forEach(function(e){r.has(e)||r.set(e,0),r.set(e,r.get(e)+1)}):(r.has(n)||r.set(n,0),r.set(n,r.get(n)+1))}),Array.from(r).map(function(e){return{key:e[0],value:e[1]}})},o=function(e,r){return arguments.length>2&&void 0!==arguments[2]&&arguments[2]?l(e,r):t.nest().key(function(e){return e[r]}).sortKeys(t.ascending).rollup(function(e){return e.length}).entries(e)};function f(r){var i=10,a=10,u=35,l=10,f="key",s="value",d=o(r.data,r.fields.y,r.fields.isList),c=(0,e.useRef)(),p=(0,e.useRef)(),y=(0,e.useRef)(),v=(0,e.useRef)(),h=function(){v.current.selectAll(".label").remove(),v.current.selectAll(".label").data(d,function(e){return e[f]}).enter().append("text").attr("class","label").attr("x",5).attr("y",function(e){return p.current.y(e[f])+p.current.y.bandwidth()/2+4}).text(function(e){return e.key}).on("click",function(e){var t={id:r.id,title:r.title,field:r.fields.y,operation:"eq",values:e.ke
[…]

[…]
