### 2024-10-20 #6837 @csoneson
@sebastian-raubach, @flekschas - could you take a look at the additional tests added above (https://github.com/sharmalab/eaglescope/pull/134) to see what you think?

@birm: regarding the community guidelines - in addition to the code of conduct (for contributors), it would be useful to have an indication in the README to guide users who would like to report an issue or seek support.
