### 2024-10-29 #6837 @birm
@csoneson -- good call. I've added a little bit to the readme under "Development" about this. https://github.com/sharmalab/eaglescope/pull/141

Thanks for the suggestion!
