### 2024-11-03 #6837 @csoneson
👋🏻 @birm - I have gone through the submission and overall it looks good! Just a couple of comments regarding the paper:

- Could you please spell out 'USA' in the affiliations?
- Also, `Emory Univeristy` should be `Emory University`
- The authors for the Bokeh references are given as "Contributors, B" - could you fix that so that it's properly displayed?

Once this is fixed, you can create a new proof with `@editorialbot generate pdf`.

I will make another post below with the steps to take after that. Thanks!
