### 2024-11-04 4b1db9b
clean up university, usa, bokeh contribs author
