+++ paper.bib
@@ -1,7 +1,7 @@
 @software{bokeh,
   title = {Bokeh: An Interactive Visualization Library for Modern Web Browsers},
   url = {https://github.com/bokeh/bokeh},
-  author = {Bokeh Contributors},
+  author = {{Bokeh Contributors}},
   date = {2024}
 }
 
