### 2024-11-04 #6837 @birm
I've put version 1.1 here: https://github.com/sharmalab/eaglescope/releases/tag/1.1
And archived here: https://zenodo.org/records/14032049

Also I've added Tony Pan's ORCID, so I'll regenerate the PDF once more:
