### 2024-11-05 #6837 @csoneson
Thank you @birm! For the Zenodo archive, could you link it directly to the GitHub repository (see [here](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content)), so that all the files are visible in the archive as well (not just the compressed repo)? Also, please make sure that the title and author list of the archive agree with those of the paper. Thanks!
