### 2024-11-05 #6837 @csoneson
Great, thanks @birm! Final detail: could you swap the order of the authors in the Zenodo archive to agree with the order in the paper? In the meanwhile I will pass the submission to the Associate EiC for the final steps.
