### 2024-11-15 #6837 @Kevin-Mattheus-Moerman
@birm as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. These are some final checks:

- [ ] Software license tag listed here matches a tagged release
